app/requirements.txt
Flask==3.0.3
google_api_python_client==2.135.0
google_auth_oauthlib==1.2.1
protobuf==5.27.2
